add more tests and improve extended contract
