+++ constants/index.ts
@@ -11,6 +11,8 @@ export default {
   CONTRACT_FACTORY_VERSION: '0.0.12',
   CONTRACT_NAME: 'MyMultiSig',
   CONTRACT_VERSION: '0.0.12',
+  CONTRACT_NAME_EXTENDED: 'MyMultiSigExtended',
   DEFAULT_THRESHOLD: 2,
-  DEFAULT_GAS: 30000,
+  DEFAULT_GAS: 50000,
+  DEFAULT_ALLOW_ONLY_OWNER: true,
 }
