+++ contracts/MyMultiSigExtended.sol
@@ -5,11 +5,11 @@ import './MyMultiSig.sol';
 
 contract MyMultiSigExtended is MyMultiSig {
   bool private _onlyOwnerRequest;
-  uint256 private _minimumTranferInactiveOwnershipAfter;
+  uint256 private _minimumTransferInactiveOwnershipAfter;
 
   struct OwnerSettings {
     uint256 lastAction;
-    uint256 tranferInactiveOwnershipAfter;
+    uint256 transferInactiveOwnershipAfter;
     address delegate;
   }
   mapping(address => OwnerSettings) private _ownerSettings;
@@ -29,6 +29,18 @@ contract MyMultiSigExtended is MyMultiSig {
   /// @return The true if the contract only accepts owner requests, false otherwise.
   function allowOnlyOwnerRequest() public view virtual returns (bool) {
     return _onlyOwnerRequest;
+  }
+
+  /// @notice Retrieves the minimum amount of time after which the other owners can transfer the ownership to a new owner
+  /// @return a uint256 representing the minimum amount of time after which the other owners can transfer the ownership to a new owner
+  function minimumTransferInactiveOwnershipAfter() public view virtual returns (uint256) {
+    return _minimumTransferInactiveOwnershipAfter;
+  }
+
+  /// @notice Retrieves owner settings
+  /// @return a OwnerSettings struct
+  function ownerSettings(address owner) public view virtual returns (OwnerSettings memory) {
+    return _ownerSettings[owner];
   }
 
   /// @notice Set if the contract only accepts owner requests (use for UI and other integrations)
@@ -72,46 +84,54 @@ contract MyMultiSigExtended is MyMultiSig {
     }
   }
 
-  /// @notice Set an amount of time after which the other owners can transfer the ownership to a new owner
-  /// @param tranferInactiveOwnershipAfter The amount of time after which the other owners can transfer the ownership to a new owner.
+  /// @notice Adds an owner
+  /// @param owner The address to be added as an owner.
   /// @dev This function can only be called inside a multisig transaction.
-  function setTranferInactiveOwnershipAftert(uint256 tranferInactiveOwnershipAfter) public virtual onlyThis {
+  function _addOwner(address owner) internal virtual override {
+    _ownersOrDelegates[owner] = true;
+    super._addOwner(owner);
+  }
+
+  /// @notice Set an amount of time after which the other owners can transfer the ownership to a new owner
+  /// @param transferInactiveOwnershipAfter The amount of time after which the other owners can transfer the ownership to a new owner.
+  /// @dev This function can only be called inside a multisig transaction.
+  function setTransferInactiveOwnershipAfter(uint256 transferInactiveOwnershipAfter) public virtual onlyThis {
     require(
-      tranferInactiveOwnershipAfter >= 7 days,
-      'MyMultiSigExtended: tranferInactiveOwnershipAfter must be greater than 7 days'
+      transferInactiveOwnershipAfter >= 7 days,
+      'MyMultiSigExtended: transferInactiveOwnershipAfter must be greater than 7 days'
     );
-    _minimumTranferInactiveOwnershipAfter = tranferInactiveOwnershipAfter;
+    _minimumTransferInactiveOwnershipAfter = transferInactiveOwnershipAfter;
   }
 
   /// @notice Owner can delegate the ownership to another address and set an amount of time after which the delegatee can take the ownership
-  /// @param tranferInactiveOwnershipAfter The amount of time after which the delegatee can take the ownership.
-  /// @param delegatee The address that will be able to take the ownership after the tranferInactiveOwnershipAfter time.
+  /// @param transferInactiveOwnershipAfter The amount of time after which the delegatee can take the ownership.
+  /// @param delegatee The address that will be able to take the ownership after the transferInactiveOwnershipAfter time.
   /// @dev This function can only be called inside a multisig transaction.
-  function setOwnerSettings(uint256 tranferInactiveOwnershipAfter, address delegatee) public virtual onlyThis {
+  function setOwnerSettings(uint256 transferInactiveOwnershipAfter, address delegatee) public virtual {
     require(
-      tranferInactiveOwnershipAfter > _minimumTranferInactiveOwnershipAfter,
-      'MyMultiSigExtended: tranferInactiveOwnershipAfter must be greater than _minimumTranferInactiveOwnershipAfter'
+      transferInactiveOwnershipAfter > _minimumTransferInactiveOwnershipAfter,
+      'MyMultiSigExtended: transferInactiveOwnershipAfter must be greater than _minimumtransferInactiveOwnershipAfter'
     );
     require(delegatee != address(0), 'MyMultiSigExtended: delegatee cannot be the zero address');
     require(!_ownersOrDelegates[delegatee], 'MyMultiSigExtended: delegatee is already an owner or delegatee');
-    _ownerSettings[msg.sender] = OwnerSettings(block.timestamp, tranferInactiveOwnershipAfter, delegatee);
+    _ownerSettings[msg.sender] = OwnerSettings(block.timestamp, transferInactiveOwnershipAfter, delegatee);
     _ownersOrDelegates[delegatee] = true;
   }
 
-  /// @notice Delegatee can take the ownership after the tranferInactiveOwnershipAfter time
+  /// @notice Delegatee can take the ownership after the transferInactiveOwnershipAfter time
   /// @param owner The owner address.
   function takeOverOwnership(address owner) external virtual {
     OwnerSettings memory ownerSettings = _ownerSettings[owner];
     require(ownerSettings.delegate == msg.sender, 'MyMultiSigExtended: msg.sender is not the delegatee');
     require(
-      ownerSettings.lastAction + ownerSettings.tranferInactiveOwnershipAfter < block.timestamp,
+      ownerSettings.lastAction + ownerSettings.transferInactiveOwnershipAfter < block.timestamp,
       'MyMultiSigExtended: owner is still active'
     );
     _ownersOrDelegates[owner] = false;
     _ownersOrDelegates[msg.sender] = true;
     _ownerSettings[msg.sender] = OwnerSettings(
       block.timestamp,
-      ownerSettings.tranferInactiveOwnershipAfter,
+      ownerSettings.transferInactiveOwnershipAfter,
       address(0)
     );
     _replaceOwner(owner, msg.sender);
