+++ test/shared/errors.ts
@@ -17,6 +17,12 @@ export default {
   NEW_OWNER_ALREADY_OWNER: 'MyMultiSig: new owner must not be an owner',
   NEW_OWNER_IS_ZERO_ADDRESS: 'MyMultiSig: new owner must not be the zero address',
 
+  INACTIVE_OWNERSHIP_MINIMUM_LESS_THAN_7DAYS:
+    'Can set an amount of time (1 day) after which the other owners can transfer the ownership',
+  OWNER_SETTINGS_MUST_BE_GREATER_THAN_MINIMUM:
+    'MyMultiSigExtended: transferInactiveOwnershipAfter must be greater than _minimumtransferInactiveOwnershipAfter',
+  OWNER_SETTINGS_DELEGATEE_MUST_NOT_BE_OWNER: 'MyMultiSigExtended: delegatee is already an owner or delegatee',
+
   PANIC_CODE_0x11:
     'VM Exception while processing transaction: reverted with panic code 0x11 (Arithmetic operation underflowed or overflowed outside of an unchecked block)',
 }
